fix lint and modify apply logic
